services/lang_detect/requirements.txt
grpcio
requests
